Fixed Coverity reported performance issues like inefficient string concatenations, wrong boxing or unboxing types, inefficent map element retrievals

+++ server/src/com/cloud/storage/VolumeApiServiceImpl.java
@@ -820,7 +820,7 @@ public class VolumeApiServiceImpl extends ManagerBase implements VolumeApiServic
         if (!shrinkOk) {
             /* Check resource limit for this account on primary storage resource */
             _resourceLimitMgr.checkResourceLimit(_accountMgr.getAccount(volume.getAccountId()), ResourceType.primary_storage, volume.isDisplayVolume(), new Long(newSize
-                    - currentSize));
+                    - currentSize).longValue());
         }
 
         /* If this volume has never been beyond allocated state, short circuit everything and simply update the database */
@@ -1667,7 +1667,7 @@ public class VolumeApiServiceImpl extends ManagerBase implements VolumeApiServic
 
                 // retrieve the migrated new volume from job result
                 if (jobResult != null && jobResult instanceof Long) {
-                    return _entityMgr.findById(VolumeVO.class, ((Long)jobResult).longValue());
+                    return _entityMgr.findById(VolumeVO.class, ((Long)jobResult));
                 }
                 return null;
             }
